Split files, cleanup
